git-svn-id: https://123.57.29.132:8443/svn/GEHCIT/APS@2493 42624c65-7419-e447-9ad7-a502218ba35c

+++ src/src/main/java/com/ge/apm/view/asset/AssetInfoController.java
@@ -228,6 +228,7 @@ public class AssetInfoController extends JpaCRUDController<AssetInfo> {
     private void saveAttachements() {
         for (AssetFileAttachment item : attachements) {
             if (null == item.getId()) {
+                item.setAssetId(selected.getId());
                 attachDao.save(item);
             }
         }
@@ -238,7 +239,9 @@ public class AssetInfoController extends JpaCRUDController<AssetInfo> {
         for (AssetFileAttachment item : attachements) {
             if (attachid.equals(item.getFileId())) {
                 fileService.deleteAttachment(item.getFileId());
-                attachDao.delete(item);
+                if(item.getId() != null){
+                    attachDao.delete(item);
+                }
                 attachements.remove(item);
                 break;
             }
@@ -329,5 +332,5 @@ public class AssetInfoController extends JpaCRUDController<AssetInfo> {
     public void setFilterLastStocktakeDate(String filterLastStocktakeDate) {
         this.filterLastStocktakeDate = filterLastStocktakeDate;
     }
-    
+
 }
